==> notebooks/01_wikipedia_processing.ipynb <==
## 위키피디아 텍스트 데이터 읽기 예시

from glob import glob

dir_list = glob('../data/text/*')
dir_list.sort()

dir_list

all_texts = []
for dir_i in dir_list:
    text_list = glob(dir_i + '/*')
    text_list.sort()
    all_texts.extend(text_list)

all_texts[:5]

len(all_texts)

with open('../data/text/AA/wiki_00') as f:
    lines = f.readlines()
    lines = [line.strip() for line in lines]

lines[:20]

def get_wiki(filename):
    with open(filename, 'r') as f:
        lines = [line.strip() for line in f.readlines()]
    return lines

wiki_example = get_wiki('../data/text/AA/wiki_00')
wiki_example[:20]

wiki_example[7]

from kss import split_sentences

sents = [sent for sent in split_sentences(wiki_example[7]) if sent]
sents

## 데이터 전처리

all_items = []

for item in wiki_example:
    sents = [sent for sent in split_sentences(item) if sent]
    all_items += sents

all_items[:20]

wiki_example[:15]

len(all_items)

import re
from kss import split_sentences

def preprocess(text):
    # 한 개의 평서문만 허용
    if text.count('.') != 1:
        return ''
    if '·' in text:
        return ''
    # \xa0
    text = re.sub('\xa0', ' ', text)

    # 괄호 안의 텍스트 제거
    text = re.sub(r'\([^)]*\)','', text)
    # <> 안의 텍스트 제거
    text = re.sub('<.+?>', '', text)
    # 큰따옴표 제거
    text = re.sub('"', '', text)
    # 다른 형태의 따옴표 처리
    text = re.sub(r'[‘’]', '', text)
    text = re.sub(r'[“”]', '', text)
    
    # 영어, 한국어를 제외한 다른 언어 제외
    if re.sub("[가-힣ㄱ-ㅎㅏ-ㅣa-zA-Z0-9.\W]", '', text):
        return ''
    
    # 너무 적은 개수의 토큰으로 이루어진 경우 제외(예: 어린 시절.)
    if len(text.split()) <= 5:
        return ''
    
    return text            

all_items[:20]

processed_text = [preprocess(x) for x in all_items if preprocess(x)]
processed_text[:20]

len(processed_text)

## 전체 데이터 정리

from glob import glob
from kss import split_sentences
import re

dir_list = glob('../data/text/*')
dir_list.sort()

all_texts = []
for dir_i in dir_list:
    text_list = glob(dir_i + '/*')
    text_list.sort()
    all_texts.extend(text_list)

print(f'전체 텍스트 파일 개수: {len(all_texts)}개\n')
print(f"['{all_texts[0]}', ..., '{all_texts[-1]}']")

def get_wiki_texts(filename):
    with open(filename, 'r') as f:
        lines = [line.strip() for line in f.readlines()]
    return lines

def preprocess(text):
    # 한 개의 평서문만 허용
    if text.count('.') != 1:
        return ''
    if '·' in text:
        return ''
    # \xa0
    text = re.sub('\xa0', ' ', text)

    # 괄호 안의 텍스트 제거
    text = re.sub(r'\([^)]*\)','', text)
    # <> 안의 텍스트 제거
    text = re.sub('<.+?>', '', text)
    # 큰따옴표 제거
    text = re.sub('"', '', text)
    # 다른 형태의 따옴표 처리
    text = re.sub(r'[‘’]', '', text)
    text = re.sub(r'[“”]', '', text)
    
    # 다른 언어 등 포함된 경우 제외
    if re.sub("[가-힣ㄱ-ㅎㅏ-ㅣa-zA-Z0-9.\W]", '', text):
        return ''
    
    # 너무 적은 개수의 토큰으로 이루어진 경우 제외(예: 어린 시절.)
    if len(text.split()) <= 5:
        return ''
    
    return text    

all_items = []

for i, text_file in enumerate(all_texts):
    text_tmp = get_wiki_texts(text_file)
    
    for item in text_tmp:
        sents = [sent for sent in split_sentences(item) if sent]
        all_items += sents  
    
    if (i + 1) % 50 == 0:
        print(f"{i+1}개 파일 완료: 총 {len(all_items)}개 문장")

print(f"\n최종: {len(all_texts)}개 파일, 총 {len(all_items)}개 문장")

len(all_items)

processed_text = [preprocess(x) for x in all_items if preprocess(x)]
processed_text[:20]

len(processed_text)

## json 파일로 저장

import json
from tqdm import tqdm

text_dict = {}

text_dict["data"] = []

for text_tmp in tqdm(processed_text):
    text_dict["data"].append({"sentence": text_tmp})

text_dict.keys()

len(text_dict['data'])

text_dict['data'][:20]

with open("../data/wikipedia/processed_sentences.json", 'w') as f:
    json.dump(text_dict, f, indent=4)

with open("../data/wikipedia/processed_sentences.json", 'r') as f:
    json_data = json.load(f)
    

json_data.keys()

len(json_data['data'])

json_data['data'][:20]
