==> notebooks/03_entity_linking.ipynb <==
## 전처리된 wikipedia corpus 불러오기

import json

with open("../data/wikipedia/processed_sentences.json", 'r') as f:
    sentences = json.load(f)

sentences.keys()

len(sentences['data'])

sentences['data'][:10]

def get_sent(idx, json_data=sentences):
    return json_data['data'][idx]['sentence']

print(get_sent(0))
print(get_sent(1))

## Pororo를 사용한 NER 수행 예시

from pororo import Pororo
ner = Pororo(task='ner', lang='ko')

ner

get_sent(0)

ner(get_sent(0))

ner(get_sent(1))

for i in range(5):
    print(ner(get_sent(i)))
    print('\n')

## Entity type 종류 확인

len(sentences['data'])

import copy

tmp = copy.deepcopy(sentences['data'][0])
tmp

ner(tmp['sentence'])

len(ner(tmp['sentence']))

from tqdm import tqdm

type_dict = {}

for idx in tqdm(range(3000)):
    ner_result = ner(get_sent(idx))
    for item in ner_result:
        if item[1] not in type_dict:
            type_dict[item[1]] = True
        

type_dict

## ent2id.json 파일 쓰기
개체명이 주어졌을 때 그 개체에 해당하는 위키데이터 id를 대응시키는 json 파일 저장

from glob import glob

file_list = glob('../data/wikidata/*')
file_list[:10]

file_list[0][26:-5]

'../data/wikidata/wikidata_1.json'[26:-5]

file_list.sort(key=lambda x: int(x[26:-5]))
file_list[:10]

from glob import glob
import json
from tqdm import tqdm

file_list = glob('../data/wikidata/*')
file_list.sort(key=lambda x: int(x[26:-5]))

result_dict = {}
for filename in tqdm(file_list):
    with open(filename, 'r') as f:
        json_data = json.load(f)
        for ent in json_data['entities']:
            try: result_dict[ent['labels']['ko']] = ent['id']
            except: pass
            
            try:
                for alias in ent['aliases']['ko']:
                    result_dict[alias] = ent['id']
            except: pass


result_dict

len(result_dict)

result_dict['제임스 얼 카터']

with open('../data/processed_data/ent2id.json', 'w', encoding='utf-8') as f:
    json.dump(result_dict, f, indent=4)

## entid2rel.json 파일 쓰기
개체 id가 주어졌을 때 그 개체가 가지는 모든 관계 정보를 대응시키는 json 파일 저장

from glob import glob
import json
from tqdm import tqdm

file_list = glob('../data/wikidata/*')
file_list.sort(key=lambda x: int(x[26:-5]))

result_dict = {}

for filename in tqdm(file_list):
    with open(filename, 'r') as f:
        json_data = json.load(f)
        
        for ent in json_data['entities']:
            tmp_rel_dict = {}
            
            for key in ent['relations']:                    
                for ent_id in ent['relations'][key]:
                    tmp_rel_dict[ent_id] = key
                    
            result_dict[ent['id']] = tmp_rel_dict


len(result_dict)

result_dict

result_dict['Q31']

with open('../data/processed_data/entid2rel.json', 'w', encoding='utf-8') as f:
    json.dump(result_dict, f, indent=4)

import sys

sys.getsizeof(result_dict)

## 위키데이터(지식 그래프) id 얻는 함수 작성

with open('../data/processed_data/ent2id.json', 'r') as f:
    ent2id = json.load(f)

def get_entity_id(entity_name: str, ent2id=ent2id) -> str:
    """ 주어진 개체명의 위키데이터 id를 얻는 함수.
    """
    return ent2id[entity_name] if entity_name in ent2id else None


get_entity_id('제임스 얼 카터')

## Entity가 가지는 관계 정보를 얻는 함수 작성

# with open('../data/processed_data/entid2rel.json', 'r') as f:
#     entid2rel = json.load(f)
    
def get_rel_info(entity_id: str, entid2rel=result_dict) -> dict:
    """ 주어진 개체 id가 가지는 관계 정보를 담은 딕셔너리를 반환하는 함수.
    """
    return entid2rel[entity_id] if entity_id in entid2rel else None

get_rel_info(get_entity_id('제임스 얼 카터'))

## 문장으로부터 관계 트리플을 얻는 함수 작성

from itertools import permutations
import random

def sent2triple(sent: str) -> list:
    """ 문장으로부터 가능한 관계 트리플의 리스트를 얻는 함수.
    
    ex) sent = '제임스 얼 카터 주니어는 민주당 출신 미국 39대 대통령 이다.'
        
        Return:
            [
                {"sentence": "제임스 얼 카터 주니어는 민주당 출신 미국 39대 대통령 이다.",
                "subj_name": "제임스 얼 카터",
                "subj_start_pos": 0,
                "subj_end_pos": 8,
                "subj_type": "PERSON",
                "obj_name": "대통령",
                "obj_start_pos": 28,
                "obj_end_pos": 31,
                "obj_type": "CIVILIZATION",
                "relation": "~",
                
                { ...
                }
            ]
    """
    # 너무 긴 문장의 경우 제외
    if len(sent) >= 500:
        return []
    
    ner_result = ner(sent)
    
    # 인식된 각 개체명의 range 계산
    ner_result = [(item[0], item[1], len(item[0])) for item in ner_result]
    
    modified_list = []
    tmp_cnt = 0
    
    for item in ner_result:
        modified_list.append((item[0], item[1], [tmp_cnt, tmp_cnt + item[2]]))
        tmp_cnt += item[2]       
    
    # NER
    ent_list = [item for item in modified_list if item[1] != 'O']
    
    # wikidata id가 존재하는 것만 선택
    ent_list = [(item[0], get_entity_id(item[0]), item[1], item[2]) for item in ent_list if get_entity_id(item[0])]

    result_list = []
    
    # 같은 문장에 대한 관계 트리플이 너무 많아지는 것을 방지하기 위해 랜덤 샘플링
    if len(ent_list) >= 6:
        sampled_pairs = random.sample(list(permutations(ent_list, 2)), 10)
    
    else:
        sampled_pairs = list(permutations(ent_list, 2))
        
    for ent_subj, ent_obj in sampled_pairs:
        tmp_dict = {}
        tmp_rel = ""
        
        if ent_obj[1] in get_rel_info(ent_subj[1]):
            tmp_rel = get_rel_info(ent_subj[1])[ent_obj[1]]
            
            if tmp_rel:
                subj_pos = ent_subj[3]
                obj_pos = ent_obj[3]
                tmp_dict["sentence"] = sent
                
                tmp_dict["subj_name"] = ent_subj[0]
                tmp_dict["subj_start_pos"], tmp_dict["subj_end_pos"] = subj_pos[0], subj_pos[1]
                tmp_dict["subj_type"] = ent_subj[2]
                
                tmp_dict["obj_name"] = ent_obj[0]
                tmp_dict["obj_start_pos"], tmp_dict["obj_end_pos"] = obj_pos[0], obj_pos[1]
                tmp_dict["obj_type"] = ent_obj[2]
                
                tmp_dict["relation"] = tmp_rel
        
                result_list.append(tmp_dict)
            
    if len(result_list) >= 3:
        sampled_result = random.sample(result_list, 2)
    else:
        sampled_result = result_list
        
    return sampled_result


sent2triple(get_sent(0))

sent2triple(get_sent(4))

## 전체 위키피디아 코퍼스의 문장에서 관계 트리플 추출하여 csv 파일로 저장

len(sentences['data'])

import random
random.shuffle(sentences['data'])
sentences['data'][0]

sentences['data'][1]

import csv
from tqdm import tqdm
from itertools import permutations

result_csv_file = open('../data/processed_data/relation_triples.csv', 'w')
writer = csv.writer(result_csv_file, delimiter='\t')
writer.writerow(['sentence', 'subj_name', 'subj_start_pos', 'subj_end_pos', 'subj_type',
                'obj_name', 'obj_start_pos', 'obj_end_pos', 'obj_type', 'relation'])

for idx, sentence_item in enumerate(tqdm(sentences['data'])):
    tmp_result_list = sent2triple(get_sent(idx))
    
    for item in tmp_result_list:
        writer.writerow([item['sentence'], item['subj_name'], item['subj_start_pos'], item['subj_end_pos'],
                        item['subj_type'], item['obj_name'], item['obj_start_pos'], item['obj_end_pos'],
                        item['obj_type'], item['relation']])
        
result_csv_file.close()

result_csv_file.close()
