==> notebooks/04_triple_data_split.ipynb <==
## 추출된 관계 트리플 데이터 불러오기

import pandas as pd

df = pd.read_csv('../data/processed_data/relation_triples.csv', delimiter='\t')

df.head()

len(df)

## 관계 종류별 개수 및 분포 비율 확인

rel_count_dict = dict(df.relation.value_counts())
list(rel_count_dict.items())[:5]

len(rel_count_dict)

rel_proportion_dict = {}

for rel in rel_count_dict:
    rel_proportion_dict[rel] = rel_count_dict[rel] / len(df) * 100

list(rel_proportion_dict.items())[:5]

list(rel_proportion_dict.items())[:100]

## wikidata의 relation 종류 가져오기 (쿼리 이용)

# pip install sparqlwrapper
# https://rdflib.github.io/sparqlwrapper/

import sys
from SPARQLWrapper import SPARQLWrapper, JSON

endpoint_url = "https://query.wikidata.org/sparql"

query = """SELECT ?property ?propertyType ?propertyLabel ?propertyDescription ?propertyAltLabel WHERE {
  ?property wikibase:propertyType ?propertyType .
  SERVICE wikibase:label { bd:serviceParam wikibase:language "[AUTO_LANGUAGE],en". }
}
ORDER BY ASC(xsd:integer(STRAFTER(STR(?property), 'P')))"""


def get_results(endpoint_url, query):
    user_agent = "WDQS-example Python/%s.%s" % (sys.version_info[0], sys.version_info[1])
    # TODO adjust user agent; see https://w.wiki/CX6
    sparql = SPARQLWrapper(endpoint_url, agent=user_agent)
    sparql.setQuery(query)
    sparql.setReturnFormat(JSON)
    return sparql.query().convert()


results = get_results(endpoint_url, query)

results_en = results["results"]["bindings"]

results_en[0]

# pip install sparqlwrapper
# https://rdflib.github.io/sparqlwrapper/

import sys
from SPARQLWrapper import SPARQLWrapper, JSON

endpoint_url = "https://query.wikidata.org/sparql"

query = """SELECT ?property ?propertyType ?propertyLabel ?propertyDescription ?propertyAltLabel WHERE {
  ?property wikibase:propertyType ?propertyType .
  SERVICE wikibase:label { bd:serviceParam wikibase:language "[AUTO_LANGUAGE],ko". }
}
ORDER BY ASC(xsd:integer(STRAFTER(STR(?property), 'P')))"""


def get_results(endpoint_url, query):
    user_agent = "WDQS-example Python/%s.%s" % (sys.version_info[0], sys.version_info[1])
    # TODO adjust user agent; see https://w.wiki/CX6
    sparql = SPARQLWrapper(endpoint_url, agent=user_agent)
    sparql.setQuery(query)
    sparql.setReturnFormat(JSON)
    return sparql.query().convert()


results = get_results(endpoint_url, query)

results_ko = results["results"]["bindings"]

results_ko[0]

## relation 정보 추출

result_dict = {}

for i, rel_item in enumerate(results_ko):
    tmp_id = rel_item['property']['value'].split('http://www.wikidata.org/entity/')[1]
    
    tmp_dict = {}
    
    tmp_dict["relation_type"] = rel_item['propertyType']['value'].split('http://wikiba.se/ontology#')[1]

    tmp_dict["relation_label"] = rel_item['propertyLabel']['value']

    try:
        tmp_dict["relation_description"] = rel_item['propertyDescription']['value']
    except:
        pass
    
    result_dict[tmp_id] = tmp_dict


list(result_dict.items())[:5]

import json

with open('../data/relation/relid2label.json', 'w', encoding='utf-8') as f:
    json.dump(result_dict, f, indent=4)

result_list = []

for i, rel_item in enumerate(results_ko):
    tmp_dict = {}
    
    tmp_dict["relation_id"] = rel_item['property']['value'].split('http://www.wikidata.org/entity/')[1]
    
    tmp_dict["relation_type"] = rel_item['propertyType']['value'].split('http://wikiba.se/ontology#')[1]

    tmp_dict["relation_label"] = rel_item['propertyLabel']['value']

    try:
        tmp_dict["relation_description"] = rel_item['propertyDescription']['value']
    except:
        pass
    
    result_list.append(tmp_dict)

result_list[:5]

rel_df = pd.DataFrame(result_list)
rel_df.head()

rel_df.tail()

results_ko[-4]

## 기존 triple 데이터에 있는 relation id에 맞게 relation_label 연결

df.rename(columns={'relation': 'relation_id'}, inplace=True)
df.head()

import numpy as np

df['relation_label'] = df['relation_id'].apply(lambda x: result_dict[x]['relation_label'])
df['relation_type'] = df['relation_id'].apply(lambda x: result_dict[x]['relation_type'])
df['relation_description'] = df['relation_id'].apply(lambda x: result_dict[x]['relation_description'] if 'relation_description' in result_dict[x] else np.nan)


df.head()

for rel_id in rel_count_dict:
    print(f"{rel_id}: {rel_count_dict[rel_id]}개, label: {result_dict[rel_id]['relation_label']}")

비율이 0.1%를 넘는 relation에 대해서만 진행

final_rel_list = [item[0] for item in list(rel_proportion_dict.items())[:97]]
print(final_rel_list[:5])
print(len(final_rel_list))

with open('../data/relation/relation_list.txt', 'w') as f:
    for rel in final_rel_list:
        f.write(rel + '\n')


with open('../data/relation/relation_list.txt', 'r') as f:
    lst = f.readlines()
lst = [x.strip() for x in lst]
lst

## 존재하는 relation 중 비율이 0.1% 이상인 97개만 선택

df['relation_id'] = df['relation_id'].apply(lambda x: x if x in final_rel_list else np.nan)
df.head()

df.info()

df.dropna(subset=['relation_id'], inplace=True)
df.head()

df.reset_index(drop=True, inplace=True)
len(df)

len(final_rel_list)

len(df), len(df[df['relation_id'].isna()])

## label에 따라 onehot label 만들기

def get_onehot(rel_id, rel_list=final_rel_list):
    onehot_list = [0] * len(rel_list)
    
    onehot_list[rel_list.index(rel_id)] = 1
    return onehot_list

final_rel_list[:5]

get_onehot('P131')

df['label_onehot'] = df['relation_id'].apply(lambda x: get_onehot(x))
df.head()

## 데이터 split

from iterstrat.ml_stratifiers import MultilabelStratifiedShuffleSplit
import numpy as np

def data_split(df, test_size=0.2):
    """ train/val/test dataset의 label의 분포 비율을 같게 split하는 함수.    
    """
    X = df.sentence.to_list()
    Y = df.label_onehot.to_list()
    
    msss = MultilabelStratifiedShuffleSplit(n_splits=2, test_size=test_size, random_state=0)
    
    for train_index, test_index in msss.split(X, Y):
        train_idx, test_idx = train_index, test_index
        
    test_df = df.iloc[test_idx.tolist()]
    tmp_df = df.iloc[train_idx.tolist()]
    
    X = tmp_df.sentence.to_list()
    Y = tmp_df.label_onehot.to_list()
    
    for train_index, test_index in msss.split(X, Y):
        train_idx, test_idx = train_index, test_index
    
    val_df = tmp_df.iloc[test_idx.tolist()]
    train_df = tmp_df.iloc[train_idx.tolist()]
    
    return train_df, val_df, test_df

train_df, val_df, test_df = data_split(df)

train_df.info()

val_df.info()

test_df.info()

train_df.head()

## 라벨별 분포 확인

from collections import Counter

def get_rel_prop(df):
    counter = Counter(df['relation_id'])
    
    count_list = [(item[0], item[1], item[1] / len(df) * 100) for item in counter.items()]

    count_list.sort(key=lambda x: x[2], reverse=True)

    return count_list

get_rel_prop(df)[:5]

get_rel_prop(train_df)[:5]

get_rel_prop(val_df)[:5]

get_rel_prop(test_df)[:5]

get_rel_prop(df)[-5:]

get_rel_prop(train_df)[-5:]

get_rel_prop(val_df)[-5:]

get_rel_prop(test_df)[-5:]

## csv 파일로 저장

import os

os.mkdir('../data/data_split')

train_df.to_csv('../data/data_split/train.csv', index=False)
val_df.to_csv('../data/data_split/val.csv', index=False)
test_df.to_csv('../data/data_split/test.csv', index=False)

## toy dataset 만들기

val_df.head()

len(val_df)

toy_df = val_df

toy_train_df, toy_val_df, toy_test_df = data_split(toy_df)

len(toy_train_df), len(toy_val_df), len(toy_test_df)

get_rel_prop(toy_train_df)[:5]

get_rel_prop(toy_val_df)[:5]

get_rel_prop(toy_test_df)[:5]

os.mkdir('../data/toy_data_split')

toy_train_df.to_csv('../data/toy_data_split/train.csv', index=False)
toy_val_df.to_csv('../data/toy_data_split/val.csv', index=False)
toy_test_df.to_csv('../data/toy_data_split/test.csv', index=False)
