==> notebooks/10_inference_from_package.ipynb <==
import sys
sys.path.append('..')

from korre import KorRE


korre = KorRE()

korre.pororo_ner('갤럭시 플립2는 삼성에서 만든 스마트폰이다.')

korre.ner('갤럭시 플립2는 삼성에서 만든 스마트폰이다.')

korre.get_all_entity_pairs('갤럭시 플립2는 삼성에서 만든 스마트폰이다.')

korre.get_all_inputs('갤럭시 플립2는 삼성에서 만든 스마트폰이다.')

korre.entity_markers_added('갤럭시 플립2는 삼성에서 만든 스마트폰이다.', [0, 7], [9, 11])

korre.infer('갤럭시 플립2는 삼성에서 만든 스마트폰이다.', [0, 7], [9, 11])

korre.infer('[E1] 갤럭시 플립2 [/E1] 는 [E2] 삼성 [/E2] 에서 만든 스마트폰이다.', entity_markers_included=True)

korre.infer('갤럭시 플립2는 삼성에서 만든 스마트폰이다.')
