==> notebooks/11_process_specific_domain_news.ipynb <==
import json
import os
from glob import glob

news_dir = '../data/newsdata'

news_list = []

for news_dir_i in glob(news_dir + '/*'):
    tmp_news_list = glob(news_dir_i + '/*')
    news_list += tmp_news_list
    
news_list = [x for x in news_list if x.endswith('.json')]
news_list.sort()

news_list

import re
from kss import split_sentences

def preprocess(text):
    # 한 개의 평서문만 허용
    if text.count('.') != 1:
        return ''
    if '·' in text:
        return ''
    # \xa0
    text = re.sub('\xa0', ' ', text)

    # 괄호 안의 텍스트 제거
    text = re.sub(r'\([^)]*\)','', text)
    # <> 안의 텍스트 제거
    text = re.sub('<.+?>', '', text)
    # 큰따옴표 제거
    text = re.sub('"', '', text)
    # 다른 형태의 따옴표 처리
    text = re.sub(r'[‘’]', '', text)
    text = re.sub(r'[“”]', '', text)
    
    # 영어, 한국어를 제외한 다른 언어 제외
    if re.sub("[가-힣ㄱ-ㅎㅏ-ㅣa-zA-Z0-9.\W]", '', text):
        return ''
    
    # 너무 적은 개수의 토큰으로 이루어진 경우 제외(예: 어린 시절.)
    if len(text.split()) <= 5:
        return ''
    
    return text      

def news2sentlist(news_file):
    ret_list = []
    with open(news_file, 'r') as f:
        json_data = json.load(f)
    
    for i in range(len(json_data['document'])):
        tmp_data = json_data['document'][i]['paragraph']
    
        for j in range(len(tmp_data)):
            if j == 0: continue
            if len(tmp_data[j]['form']) <= 70:
                ret_list.append(tmp_data[j]['form'])
                
    ret_list = [preprocess(x) for x in ret_list if preprocess(x)]
    
    return ret_list

default_dir = '../data/newsdata/news_1_1/'
len(news2sentlist(default_dir + 'NPRW2000000001.json'))

domain_sent = news2sentlist(default_dir + 'NPRW2000000001.json')

import pandas as pd

domain_df = pd.DataFrame(domain_sent)
domain_df.head()

domain_df.to_csv('../data/specific_domain_sentences.csv', index=False)


domain_sent[:5]

## entity 추출

from itertools import permutations
from pororo import Pororo

ner = Pororo(task='ner', lang='ko')


import requests

r = requests.get('https://raw.githubusercontent.com/datawhales/Korean_RE/main/data/relation/relid2label.json')
relid2label = json.loads(r.text)

list(relid2label.items())[:5]

relation_list = list(relid2label.keys())
relation_list[:5]

def idx2relid(idx_list):
    label_out = []
    
    for idx in idx_list:
        label = relation_list[idx]
        label_out.append(label)
        
    return label_out

def ner_list(sentence):
    ner_result = ner(sentence)
    ner_result = [(item[0], item[1], len(item[0])) for item in ner_result]
    
    modified_list = []
    tmp_cnt = 0

    for item in ner_result:
        modified_list.append((item[0], item[1], [tmp_cnt, tmp_cnt + item[2]]))
        tmp_cnt += item[2]
    
    true_list = {'PERSON': True,'ORGANIZATION': True, 'COUNTRY': True, 'LOCATION': True, 'EVENT': True,
                 'THEORY': True, 'ARTIFACT': True, 'ANIMAL': True, 'TERM': True, 'STUDY_FIELD': True,
                 'MATERIAL': True, 'CITY': True, 'DISEASE': True}
    ent_list = [item for item in modified_list if item[1] in true_list]

    return ent_list

for i in range(len(domain_sent)):
    print(ner_list(domain_sent[i]))
    if i == 10: break

import random

def get_all_entity_pairs(sentence):
    
    ent_list = ner_list(sentence)
    
#     if len(ent_list) >= 5:
        
#         return random.sample(list(permutations(ent_list, 2)), 3)
    
#     elif len(ent_list) <= 2:
#         return list(permutations(ent_list, 2))
    
#     return random.sample(list(permutations(ent_list, 2)), 3)
    return list(permutations(ent_list, 2))
    
def get_all_inputs_csv_form(sentence):
    pairs = get_all_entity_pairs(sentence)
    return [[sentence, sentence[ent_subj[2][0]:ent_subj[2][1]], ent_subj[2][0], 
             ent_subj[2][1], ent_subj[1], sentence[ent_obj[2][0]:ent_obj[2][1]],
             ent_obj[2][0], ent_obj[2][1], ent_obj[1]] for ent_subj, ent_obj in pairs]

get_all_entity_pairs(domain_sent[0])

get_all_inputs_csv_form(domain_sent[0])

from tqdm import tqdm

def get_news_test_data(sentences):
    ret_df = pd.DataFrame(columns=['sentence', 'subj_name', 'subj_start_pos', 'subj_end_pos',
                         'subj_type', 'obj_name', 'obj_start_pos', 'obj_end_pos', 'obj_type'])
    for i, sent in enumerate(tqdm(sentences)):
        tmp_df = pd.DataFrame(get_all_inputs_csv_form(sent),
                    columns=['sentence', 'subj_name', 'subj_start_pos', 'subj_end_pos',
                         'subj_type', 'obj_name', 'obj_start_pos', 'obj_end_pos', 'obj_type'])
        ret_df = pd.concat([ret_df, tmp_df], ignore_index=True)
#         if i == 10: break
    return ret_df

domain_test_df = get_news_test_data(domain_sent)

domain_test_df.head()

domain_test_df.to_csv('../data/news_processed/specific_domain_test.csv', index=False)

import pandas as pd

new = pd.read_csv('../data/news_processed/specific_domain_test.csv')
new.head()

new.tail()

new.sample(5)


len(new)
