==> notebooks/14_specific_domain_results.ipynb <==
import pandas as pd
import numpy as np

df = pd.read_csv('../results/specific_domain_pred_results.csv')

df.head()

len(df)

df.info()

true_df = df.dropna(subset=['relation'])
true_df

true_df.to_csv('../results/specific_domain_pred_true.csv', index=False)

import pandas as pd

new = pd.read_csv('../results/specific_domain_pred_true.csv')

new.head()

new.sample(5)

new.sample(5)

new.sample(5)
