==> notebooks/15_relation_triples_to_neo4j.ipynb <==
## neo4j 사용 예시

!pip install neo4j

from neo4j import GraphDatabase
import pandas as pd

host = 'bolt://100.24.23.207:7687'
user = 'neo4j'
password = 'patient-tolerances-grams'
driver = GraphDatabase.driver(host,auth=(user, password))

def neo4j_query(query, params=None):
    with driver.session() as session:
        result = session.run(query, params)
        return pd.DataFrame([r.values() for r in result], columns=result.keys())

author = 'ABC'
title = 'ABC books'

neo4j_query("""
MERGE (a:Author{name:$author})
MERGE (b:Article{title:$title})
MERGE (a)-[:WROTE]->(b)
""", {'title':title, 'author':author})

## triple2graph

def reltriple2graph(subj_name, subj_type, obj_name, obj_type, relation):
    query_string = 'MERGE (a:' + subj_type + '{label:$subj_name})' + \
                      'MERGE (b:' + obj_type + '{label:$obj_name})' + \
                      'MERGE (a)-[:' + relation + ']->(b)'
    query_info_dict = {'subj_name': subj_name, 'obj_name': obj_name}
    
    neo4j_query(query_string, query_info_dict)
    return query_string

subj_name = '갤럭시 플립2'
subj_type = 'ARTIFACT'
obj_name = '삼성전자'
obj_type = 'ORGANIZATION'
relation = '해당_제품의_제조사'

reltriple2graph(subj_name, subj_type, obj_name, obj_type, relation)

## 추출된 triple data 전처리

import pandas as pd

df = pd.read_csv('../results/specific_domain_pred_true.csv')

df.head()

df.relation.value_counts()

import re

def convert_relation(rel_string):
    """
    Example:
        rel_string = '다음 국가의 것임(country)'

    Return:
        '다음_국가의_것임'
    """
    # 괄호 제거
    rel_string = re.sub(r'\([^)]*\)','', rel_string)
    
    rel_string = re.sub('·', '_', rel_string)
    
    string_list = rel_string.split()
    
    return '_'.join(string_list)

convert_relation('다음 국가의 것임(country)')

convert_relation('해당 개체의 제품(product or material produced)')

df['converted_relation'] = df['relation'].apply(lambda x: convert_relation(x))

df.head()

from tqdm import tqdm

for i in tqdm(range(len(df))):
    tmp_row = df.iloc[i]
    
    subj_name, obj_name = tmp_row['subj_name'], tmp_row['subj_type']
    obj_name, obj_type = tmp_row['obj_name'], tmp_row['obj_type']
    relation = tmp_row['converted_relation']
    
    reltriple2graph(subj_name, subj_type, obj_name, obj_type, relation)
